requirements.txt
python-dotenv
discord
PyNaCl
youtube_dl
spotdl